backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-google-genai==2.0.10
google-generativeai==0.8.4
openai
chromadb
pydantic
python-dotenv
pypdf
docx2txt
flask
flask-cors
faiss-cpu
pymupdf